feat: Enhance BulkDeployModal and related components for improved deployment handling and state management
- Refactored BulkDeployModal to streamline image selection and deployment logic.
- Integrated new utility functions for better handling of deployment window metadata and bulk CD details.
- Updated service functions to fetch deployment window maps and handle API responses more effectively.
- Improved type definitions for better clarity and maintainability.
- Adjusted BulkTriggerSidebar to reflect changes in error handling for unauthorized access.
- Cleaned up unused constants and imports across the affected files.
